Separate library from binary
