+++ src/lib.rs
@@ -1,5 +1,3 @@
-#[macro_use]
-extern crate clap;
 extern crate env_logger;
 extern crate lettre;
 #[macro_use]
@@ -8,7 +6,6 @@ extern crate native_tls;
 extern crate rayon;
 extern crate trust_dns_resolver;
 
-use clap::App;
 use lettre::smtp::{SMTP_PORT, SUBMISSIONS_PORT, SUBMISSION_PORT};
 use rayon::prelude::*;
 use std::process;
@@ -16,18 +13,8 @@ use std::process;
 mod mx_hosts;
 mod smtp;
 
-fn main() {
-	env_logger::init();
-
-	// The YAML file is found relative to the current file, similar to how modules are found
-	let yaml = load_yaml!("cli.yml");
-	let matches = App::from_yaml(yaml).get_matches();
-
-	let from_email = matches.value_of("from").unwrap_or("user@example.org");
-	// Calling .unwrap() is safe here because "TO" is required
-	let to_email = matches.value_of("TO").unwrap();
-
-	debug!("User inputted email '{}'", to_email);
+pub fn email_exists(from_email: &str, to_email: &str) -> bool {
+	debug!("Checking email '{}'", to_email);
 
 	let domain = match to_email.split("@").skip(1).next() {
 		Some(i) => i,
@@ -49,7 +36,7 @@ fn main() {
 		}
 	}
 
-	let found = combinations
+	combinations
 		.par_iter() // Parallelize the find_any
 		.find_any(
 			|(host, port)| match smtp::email_exists(from_email, to_email, host, *port) {
@@ -59,7 +46,5 @@ fn main() {
 				}
 				_ => false,
 			},
-		).is_some();
-
-	println!("{}", found);
+		).is_some()
 }
